Merge pull request #3 from Mark-Hopkins-at-Williams/work

+++ nllb_files/helsinki_train.py
@@ -10,13 +10,22 @@ from tqdm import tqdm
 from datetime import datetime
 from finetune import tokenize
 from validate import evaluate_translations, log_evaluation, batched_translate
-from finetune import finetune
-from multilingualdata import MultilingualCorpus, MixtureOfBitexts, StreamingBitext, Bitext
+from finetune import finetune, cleanup, tokenizer_texts
+from multilingualdata import MultilingualCorpus, MixtureOfBitexts, StreamingBitext, Bitext, StreamingMonolingual, Monolingual
 from transformers.optimization import Adafactor
 from transformers import get_constant_schedule_with_warmup
 from transformers import AutoModelForSeq2SeqLM, AutoTokenizer, AutoConfig
-from configure import ALL_AMERICAS_NLP_LANGS, USE_CUDA
-import shutil  # for deleting directories
+from configure import *
+import shutil
+
+def flatten_sents(sents):
+    return [item for sublist in sents for item in (sublist if isinstance(sublist, list) else [sublist])]
+    # Ensure sents is a list of strings, and filter out non-string or empty sentences
+    if isinstance(sents, list):
+        sents = [str(sent) for sent in sents if isinstance(sent, str) and len(sent) > 0]
+    else:
+        sents = [str(sents)]  # If it's not a list, make it one
+    return sents
 
 def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_model_dir,
                   training_steps=60000,
@@ -35,6 +44,7 @@ def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_m
     
     # Load model and tokenizer
     if isinstance(base_model, list):  # handle pretrained model input as list
+        print('loading model and tokenizer passed in as argument')
         model = base_model[0]
         tokenizer = base_model[1]
     else:
@@ -60,12 +70,15 @@ def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_m
 
     # Initialize training
     last_best = 0
-    patience = 80000
+    patience = training_steps * 0.5
     train_losses = []
     best_models = []  # List to store best models and their scores
 
     for i in tqdm(range(training_steps)):
         lang1_sents, lang2_sents, lang1, lang2 = mixture_of_bitexts.next_batch()
+        
+        if lang1 == '<gen>':
+            lang1_sents = lang1_sents[0]
         try:
             model.train()
             x = tokenize(lang1_sents, lang1, tokenizer, max_length).to(model.device)
@@ -92,13 +105,14 @@ def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_m
             print("Validating on a sample...")
             sys.stdout.flush()
             avg_chrf = 0
-            num_val_lang = math.ceil(len(dev_bitext_list) / 3)
+            num_val_lang = math.ceil(len(dev_bitext_list))  ### changed this to eval all langs at each eval
             print("# evals at this validation:", num_val_lang)
             sys.stdout.flush()
-            for _ in range(num_val_lang):
-                bitext_index = random.choices(range(len(dev_bitext_list)), weights=sampling_probs, k=1)[0]
+            for q in range(num_val_lang):
+                bitext_index = q
                 dev_bitext = dev_bitext_list[bitext_index]
-                if dev_bitext.lang1_code == "eng_Latn" or dev_bitext.lang2_code == "eng_Latn":
+                print(dev_bitext.lang2_code)
+                if dev_bitext.lang1_code == "eng_Latn" or dev_bitext.lang2_code == "eng_Latn": 
                     src_texts = []
                     tgt_texts = []
                     for _ in range(100):
@@ -108,6 +122,7 @@ def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_m
                         assert len(src_texts) == len(tgt_texts)
                 else:
                     src_texts, tgt_texts = dev_bitext.lang1_sents, dev_bitext.lang2_sents
+                    
                 candidate_translations = batched_translate(src_texts, tokenizer=tokenizer, model=model, src_lang=dev_bitext.lang1_code, tgt_lang=dev_bitext.lang2_code)
                 bleu, chrf = evaluate_translations(candidate_translations, tgt_texts)
                 avg_chrf += chrf
@@ -129,7 +144,7 @@ def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_m
                 best_models.append((avg_chrf, model_dir))
                 
                 # Keep only the top `num_best_save` models
-                if len(best_models) > num_best_save:
+                if len(best_models) > num_best_save and num_best_save != 1:
                     # Remove the model with the lowest score
                     worst_model = min(best_models, key=lambda x: x[0])
                     shutil.rmtree(worst_model[1])  # Delete the model directory
@@ -140,12 +155,6 @@ def finetune_n_best(mixture_of_bitexts, dev_bitext_list, base_model, finetuned_m
             # Update last best tracking
             if avg_chrf > last_best:
                 last_best = i
-
-        # Check for patience
-        if i - last_best >= patience:
-            print('No improvement in', patience, 'epochs. Stopping training.')
-            sys.stdout.flush()
-            break
     
     # Return the directories of the saved models
     best_model_dirs = [model_dir for _, model_dir in sorted(best_models, reverse=True)]
@@ -159,13 +168,12 @@ def helsinki_bitexts_list(opus_corpus, opus_files, anlp_corpus, split):
     bitext_list = []
     
     # append eng_spa
-    bitext_list.append(opus_corpus.create_bitext('eng_Latn', 'spa_Latn', split, lang1_file = opus_files[0], lang2_file = opus_files[1]))
+    bitext_list.append(opus_corpus.create_bitext('spa_Latn', 'eng_Latn', split, lang1_file = opus_files[0], lang2_file = opus_files[1]))
     # append all anlp _ spa
-    for lang in ALL_AMERICAS_NLP_LANGS:
+    for lang in ALL_LANGS:
         bitext_list.append(anlp_corpus.create_bitext('spa_Latn', lang, split))
     return bitext_list
-
-    
+  
 if __name__ == "__main__":
     
     parser = argparse.ArgumentParser(description="Three-phase training for multilingual translation")
@@ -177,20 +185,13 @@ if __name__ == "__main__":
     print(f"Starting Training Script for {save_dir}")
     sys.stdout.flush()
     
-    # Load in model and tokenizer from scratch
-    print("Loading in model for training")
-    sys.stdout.flush()
-    base_model = "facebook/nllb-200-distilled-600M"
-    tokenizer = AutoTokenizer.from_pretrained(base_model)
-    config = AutoConfig.from_pretrained(base_model)
-    model_base = AutoModelForSeq2SeqLM.from_config(config)
     
-    # load in data 
-    anlp_csv = "americas_nlp_data.csv"
+    ######## DATA 
+    anlp_csv = "data/americas_nlp_data.csv"
     opus_csv = None
-    opus_dir = "opus_data"
-    opus_files_train = (f'{opus_dir}/train.en', f'{opus_dir}/train.es')
-    opus_files_dev = (f'{opus_dir}/dev.en', f'{opus_dir}/dev.es')
+    opus_dir = "data/opus_data"
+    opus_files_train = (f'{opus_dir}/train.es', f'{opus_dir}/train.en')
+    opus_files_dev = (f'{opus_dir}/dev.es', f'{opus_dir}/dev.en')
     
     print("Loading in data")
     sys.stdout.flush()
@@ -201,13 +202,30 @@ if __name__ == "__main__":
     dev_bitexts = helsinki_bitexts_list(opus_corpus, opus_files_dev, anlp_corpus, 'dev')
     # create all bitexts
     
+    '''
+    ###### MODEL AND TOKENIZER 
+    # Load in model and tokenizer from scratch
+    print("Loading in model for training")
+    sys.stdout.flush()
+    base_model = "facebook/nllb-200-distilled-600M"
+    tokenizer = AutoTokenizer.from_pretrained(base_model)
+    
+    #TODO: create the iterator that evenly samples each langauge to train tokenizer 
+    sents = tokenizer_texts(opus_corpus, opus_files_train, anlp_corpus, 10000) 
+    tokenizer = tokenizer.train_new_from_iterator(sents, 100000)
+    config = AutoConfig.from_pretrained(base_model)
+    model_base = AutoModelForSeq2SeqLM.from_config(config)
+    
+    
     # create phase 1 data 91% eng 9% indigenous
     print('Making phase 1 data')
     sys.stdout.flush()
+    
+    ##### PHASE 1
     phase1_p = make_sampling_weights(0.91, 0.09)
-    phase1_data = MixtureOfBitexts(train_bitexts, batch_size = 2, sampling_probs = phase1_p)
-    phase1_dev_data = [opus_corpus.create_bitext('eng_Latn', 'spa_Latn', 'dev', lang1_file = opus_files_dev[0], lang2_file = opus_files_dev[1])]
-    phase1_dir = save_dir + 'phase_1'
+    phase1_data = MixtureOfBitexts(train_bitexts, batch_size = 16, sampling_probs = phase1_p)
+    phase1_dev_data = [opus_corpus.create_bitext('spa_Latn', 'eng_Latn', 'dev', lang1_file = opus_files_dev[0], lang2_file = opus_files_dev[1])]
+    phase1_dir = save_dir + '/phase_1'
         
     # train phase 1 
     print("Starting phase 1 training")
@@ -215,19 +233,23 @@ if __name__ == "__main__":
     phase1_model_tok = [model_base, tokenizer]
     phase1_model, phase1_tokenizer = finetune(phase1_data, phase1_dev_data, phase1_model_tok, phase1_dir, training_steps=100000)
     print(f'Done training phase 1, saved to {phase1_dir}. Now making phase 2 data')
-    sys.stdout.flush()
+    sys.stdout.flush()'''
         
+    ##### PHASE 2
     # create phase 2 data 37% eng 63% indegenous
+    base_model = "Helsinki-NLP/opus-mt-es-en"
+    phase1_model = AutoModelForSeq2SeqLM.from_pretrained(base_model)
+    phase1_tokenizer = AutoTokenizer.from_pretrained(base_model)
+    
     phase2_p = make_sampling_weights(0.37, 0.63)
-    phase2_data = MixtureOfBitexts(train_bitexts, batch_size = 2, sampling_probs = phase2_p)
+    phase2_data = MixtureOfBitexts(train_bitexts, batch_size = 16, sampling_probs = phase2_p)
     phase2_dev_data = dev_bitexts
-    phase2_dir = save_dir + 'phase_2'
+    phase2_dir = save_dir + '/phase_2'
     
     # train phase 2
     print("Starting phase 2 training")
     sys.stdout.flush()
     phase2_model_tok = [phase1_model, phase1_tokenizer]
-    phase2_models = finetune_n_best(phase2_data, phase2_dev_data, phase2_model_tok, phase2_dir, training_steps=200000, sampling_probs = phase2_p, num_best_save = 5)
+    phase2_models = finetune_n_best(phase2_data, phase2_dev_data, phase2_model_tok, phase2_dir, training_steps=200000, sampling_probs = phase2_p, num_best_save = 5, validate_every=5000)
     print(f"best models from phase 2: {phase2_models}")
     sys.stdout.flush()
-
